修复侧键返回的触发
